basic validation and mailing service

+++ routes/courses.js
@@ -3,39 +3,70 @@ const Course = require('../models/course');
 const router = Router();
 const auth = require('../middleware/auth');
 
-router.get('/', async (req, res) => {
-  const courses = await Course.find()
-    .populate('userId', 'email name')
-    .select('price title img')
-    .lean();
+function isOwner(course, req) {
+  return course.userId.toString() === req.user._id.toString();
+}
 
-  res.render('courses', {
-    title: 'Курсы',
-    isCourses: true,
-    courses
-  });
+router.get('/', async (req, res) => {
+  try {
+    const courses = await Course.find()
+      .populate('userId', 'email name')
+      .select('price title img')
+      .lean();
+
+    res.render('courses', {
+      title: 'Курсы',
+      isCourses: true,
+      userId: req.user ? req.user._id.toString() : null,
+      courses,
+    });
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 router.get('/:id/edit', auth, async (req, res) => {
   if (!req.query.allow) {
     return res.redirect('/');
   }
-  const course = await Course.findById(req.params.id).lean();
-  res.render('course-edit', { title: `Редактировать ${course.title}`, course });
+
+  try {
+    const course = await Course.findById(req.params.id).lean();
+    if (!isOwner(course, req)) {
+      return res.redirect('/courses');
+    }
+
+    res.render('course-edit', {
+      title: `Редактировать ${course.title}`,
+      course,
+    });
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 router.post('/edit', auth, async (req, res) => {
-  const { id } = req.body;
-  delete req.body.id;
+  try {
+    const { id } = req.body;
+    delete req.body.id;
+    const course = await Course.findById(id);
 
-  await Course.findByIdAndUpdate(id, req.body);
-  res.redirect('/courses');
+    if (!isOwner(course, req)) {
+      return res.redirect('/courses');
+    }
+
+    Object.assign(course, req.body);
+    await course.save();
+    res.redirect('/courses');
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 router.post('/remove', auth, async (req, res) => {
   const { id } = req.body;
   try {
-    await Course.deleteOne({ _id: id });
+    await Course.deleteOne({ _id: id, userId: req.user._id });
     res.redirect('/courses');
   } catch (e) {
     console.log(e);
@@ -43,13 +74,17 @@ router.post('/remove', auth, async (req, res) => {
 });
 
 router.get('/:id', async (req, res) => {
-  const course = await Course.findById(req.params.id).lean();
+  try {
+    const course = await Course.findById(req.params.id).lean();
 
-  res.render('course', {
-    layout: 'empty',
-    title: `Курс ${course.title}`,
-    course
-  });
+    res.render('course', {
+      layout: 'empty',
+      title: `Курс ${course.title}`,
+      course,
+    });
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 module.exports = router;
